backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
lxml==6.0.2
openpyxl==3.1.5
pillow==12.1.1
python-docx==1.2.0
sqlparse==0.5.4
typing_extensions==4.15.0
